deploy: factor out websocket connection to a separate module
